inha-macos
Eu te amo Carlos Gustavo Cecyn Lundgren!!

+++ os/macos/operational_system/cross/win/win_def.h
@@ -314,20 +314,20 @@ typedef const RECTL FAR* LPCRECTL;
 
 typedef struct tagPOINT
 {
-    LONG  x;
-    LONG  y;
+    int  x;
+    int  y;
 } POINT, *PPOINT, NEAR *NPPOINT, FAR *LPPOINT;
 
 typedef struct _POINTL      /* ptl  */
 {
-    LONG  x;
-    LONG  y;
+    int  x;
+    int  y;
 } POINTL, *PPOINTL;
 
 typedef struct tagSIZE
 {
-    LONG        cx;
-    LONG        cy;
+    int        cx;
+    int        cy;
 } SIZE, *PSIZE, *LPSIZE;
 
 typedef SIZE               SIZEL;
